refactor(vite-node-miniflare): organize code
