+++ packages/vite-node-miniflare/package.json
@@ -28,7 +28,6 @@
   "dependencies": {
     "@hattip/adapter-node": "^0.0.34",
     "@hattip/compose": "^0.0.34",
-    "esbuild": "^0.18.6",
     "miniflare": "^3.20231030.3"
   },
   "devDependencies": {
@@ -37,13 +36,7 @@
     "vite": "^5.1.0"
   },
   "peerDependencies": {
-    "esbuild": "*",
     "miniflare": "*",
     "vite": "*"
-  },
-  "peerDependenciesMeta": {
-    "esbuild": {
-      "optional": true
-    }
   }
 }
