+++ packages/vite-node-miniflare/src/index.ts
@@ -1,2 +1,2 @@
 export { vitePluginViteNodeMiniflare } from "./server/plugin";
-export type { ViteNodeMiniflareClient } from "./client/vite-node";
+export type { ViteNodeMiniflareClient } from "./client/worker-entry";
